requirements.txt
# dislord~=0.0.1
discord-interactions~=0.4.0
requests~=2.31.0
python-dotenv==1.0.0
pynamodb==6.0.1
fastapi==0.115.0
